+++ src/main/java/roomescape/repository/ThemeDao.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Repository;
 import roomescape.domain.Theme;
 
 import java.sql.PreparedStatement;
+import java.time.LocalDate;
 import java.util.Collections;
 import java.util.List;
 import java.util.Optional;
@@ -18,6 +19,13 @@ public class ThemeDao {
 
     private final JdbcTemplate jdbcTemplate;
     private final RowMapper<Theme> timeRowMapper = (resultSet, rowNum) -> new Theme(
+            resultSet.getLong("id"),
+            resultSet.getString("name"),
+            resultSet.getString("description"),
+            resultSet.getString("thumbnail")
+    );
+
+    private RowMapper<Theme> themeRowMapper = (resultSet, rowNum) -> new Theme(
             resultSet.getLong("id"),
             resultSet.getString("name"),
             resultSet.getString("description"),
@@ -59,5 +67,24 @@ public class ThemeDao {
 
     public void deleteById(Long id) {
         jdbcTemplate.update("DELETE FROM theme WHERE id = ?", id);
+    }
+
+    public List<Theme> findThemesByDescOrder() {
+        String nowDate = LocalDate.now().toString();
+        String weekBeforeDate = LocalDate.now().minusDays(7).toString();
+        return jdbcTemplate.query("""
+                SELECT
+                th.id as theme_id,
+                th.name,
+                th.description,
+                th.thumbnail,
+                COUNT(r.theme_id) AS reservation_count
+                FROM theme AS th
+                INNER JOIN reservation AS r ON r.theme_id = th.id
+                WHERE r.date < ? AND r.date > ?
+                GROUP BY th.id, th.name, th.description, th.thumbnail
+                ORDER BY reservation_count DESC
+                LIMIT 10
+                """, themeRowMapper, nowDate, weekBeforeDate);
     }
 }
